:recycle: extract Bar component
