+++ src/components/navigations/global-navigation-bar-small/index.tsx
@@ -1,7 +1,7 @@
+import Bar from '@components/navigations/global-navigation-bar-small/Bar';
 import NavigationTitle from '@components/navigations/global-navigation-bar-small/NavigationTitle';
 import Notification from '@components/navigations/global-navigation-bar-small/Notification';
 import { css, styled } from '@mui/material';
-import { MLAB_OPACITY_PALETTE } from '@styles/mlabTheme';
 import Menu from './Menu';
 import Search from './Search';
 
@@ -12,21 +12,6 @@ interface GlobalNavigationBarSmallProps {
     isLogged: boolean;
 }
 
-const BAR_HEIGHT = '3.5rem';
-
-const Bar = styled('header')(({ theme }) => css`
-  height: ${BAR_HEIGHT};
-  background-color: ${theme.palette.mode === 'dark' ? MLAB_OPACITY_PALETTE.black[600] : MLAB_OPACITY_PALETTE.white[600]};
-  backdrop-filter: blur(30px);
-  position: fixed;
-  width: 100%;
-  z-index: ${theme.zIndex.appBar};
-  border-bottom: 1px solid ${theme.palette.mode === 'dark' ? theme.palette.gray[900] : theme.palette.gray[200]};
-  padding: 0 1rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`);
 
 const NavigationButtons = styled('div')(({ theme }) => css`
   display: flex;
